Return instead of await requests
https://eaflood.atlassian.net/browse/WATER-4375

> Part of a series of changes related to replacing the create bill run journey to incorporate changes for two-part tariff

The code that calls these requests should have the option to decide whether the request needs to be awaited or not. If we return directly rather than `const results await` they have that option.

This updates the requests to return rather than await the result. There will be no impact on the code because it is already awaiting these requests when they are used.