Add first route to bill run setup journey
https://eaflood.atlassian.net/browse/WATER-4375

> Part of a series of changes related to replacing the create bill run journey to incorporate changes for two-part tariff

First thing to note is we are referring to the journey to create a bill run as setting up a bill run i.e. 'setup'. Though the big green button on the UI page says 'Create' we consider creating a bill run to be more the actual creation of the bill run record and then all the subsequent bills and transactions.

Before you can get there you have to 'setup' what kind of bill run you want create. Hence, our name for choice for the path that will sit under bill runs; `/bill-runs/setup/`.

This change is the first route in the journey. This is where we will create the session record that will be used to record the choices a user makes when setting up a bill run. It will then redirect to the first page in the journey (coming in a subsequent change!)