Add bill runs setup region page to journey
https://eaflood.atlassian.net/browse/WATER-4375

> Part of a series of changes related to replacing the create bill run journey to incorporate changes for two-part tariff

This adds the second page to the setup bill run journey; select a region. This has a bit more complexity than [Add bill runs setup type page to journey](https://github.com/DEFRA/water-abstraction-system/pull/802) because depending on the bill run type selected, we are either done or need to go to another page in the journey.