Handle bill run setup matches an existing bill run
https://eaflood.atlassian.net/browse/WATER-4375

> Part of a series of changes related to replacing the create bill run journey to incorporate changes for two-part tariff

At the end of the bill run setup journey, before we attempt to generate the bill run we first have to check if an existing bill run matches the options selected during setup.

The criteria for 'matching' differs depending on bill run type.

- **Annual** - does an annual bill run for the same region and current financial year exist?
- **Supplementary** - does an supplementary bill run for the same region with a status of `BUILDING` or `READY` exist?
- **Two-part tariff (SROC)** - does a 2PT bill run for the same region and selected financial year exist?
- **Two-part tariff (PRESROC)** - does a 2PT bill run for the same region, selected financial year and season exist?

> in all cases we exclude those that have a status of `CANCEL`, `EMPTY` or `ERROR`

It is further complicated by the message we need to show when a match is found. For example, if there is a matching annual bill run with a status of `SENT`, then the message is a flat 'You cannot do this'. If however the matching bill run has a status of `READY` the message becomes 'You have to delete this first'.

This change adds the logic that will do that checking plus the view to display the result.