Submit Check your answers page
https://eaflood.atlassian.net/browse/WATER-4264

This page allows the user to submit their answers.

Ther page will run submit the data to the check your answers service which in turn will check the licence.

If the licence is valid, the user will be directed to the approved page.

If the licence is invalid, the service will throw an error.