Persist and retrieve rtn req set up session values
https://eaflood.atlassian.net/browse/WATER-4384

This PR will allow the application to persist data through the 'return-requirements' journey.

The data will persist in the session until it is cleared by another part of the application.

The data will be used to populate controls and fields so that the user can edit or review, and also for UI updates.