Fix ExistsService yearToUse
Spotted in tests that when the service determined that the year to use was what the user selected it was returning it as a string. However, for supplementary and annual where it uses the current financial year it was returning a number.

This inconsistency was causing a problem for the CreateService (and just is wrong!)