Fix failing fetch-licence.service.test.js
The fetch-licence.service.test.js was failing because it was getting a stubbed result back when what we wanted was the actual service to be called.

The reason it was still stubbed out was because submit-check-your-answers.test.js was stubbing the service outside of a `beforeEach()`. This is the equivalent of stubbing the service as soon as the test suite is loaded (which it always is even if you then just run one test).

It's outside the scope of the tests and therefore a `Sinon.restore()` in an `afterEach()` is not going to apply.

Putting it inside of a `beforeEach()` means the stubbing now only happens in the scope of the `submit-check-your-answers.service.test.js`.