Housekeeping in FetchLicenceService
When dealing with async code unless we need to await the result (as we do in `go()` in order to be able to pass the licence to `_data()`) we should be returning immediately so the calling code then has the option to await or not.

The other cleanup is that `_data()` was not doing anything blocking so didn't need to be `async`.