Fix supplementary matching logic in bill run setup
https://eaflood.atlassian.net/browse/WATER-4375

> Part of a series of changes related to replacing the create bill run journey to incorporate changes for two-part tariff

Having completed the migration of the setup bill run journey we deployed to our test environment and did some initial checking.

In that environment **Wales** had a two-part tariff bill run in a state of `READY` but nothing else. So, when trying to create a supplementary bill run we expected to see both an SROC and PRESROC bill run created.

However, only the SROC one was created. Debugging confirmed we were sending the request to the legacy service but it was returning a `409 - conflict`. Further checking confirmed that it was seeing the two-part tariff as an existing bill run that should block the supplementary from being processed.

"Well legacy is wrong!" was our immediate response. But on reflection it makes perfect sense. The idea of a supplementary is to recalculate charges to a customer because a change has been made. But it can only do that if the existing bill runs it is looking at are also not still being processed.

So, if the user is attempting to setup a supplementary bill run, _any_ in progress bill runs irrespective of type should block it.

This change updates the `ExistsService` and `FetchMatchingBillRunService` to cater for this new understanding.

We also put some logging around the request to the legacy request so we can more easily see issues like this in the future.