Add new FetchLiveBillRuns service
Having double checked the legacy code we now understand the check when creating a new bill run is done in 2 parts.

- is there _any_ live bill runs for the matching year?
- is there any duplicate bill runs?

In the case of supplementary it just does the live check. For annual and two-part tariff it does both. The takeaway is if _any_ bill run is in progress (live) for the year being processed then we are expected to block generation from taking place.

Our job gets a little more complex because unlike the legacy service, when it comes to supplementary we have to cater for there possibly being two; SROC and PRESROC. So, our service takes an extra flag which tells it if we are dealing with a supplementary bill run. If we are we not only do we check the financial end year provided, but any live bill runs in 2022 which is the last year of PRESROC.

> When you request the legacy service to create a supplementary bill run the request will include the current financial year's end date as the `financialEndYear`. The legacy service will then adjust this to be the last year of PRESROC; 2022. So, we know when it comes to supplementary we are always checking for matches in the current financial year and 2022.