Add new BillRunMatchService
This is the result of starting to hack the changes needed in a temporary service. The intent was to refactor `ExistsService` to incorporate the changes. But once we got to the end it was clear we'd need something separate.

So, we cleaned up the hack and this new service is the result.