Add relationship between bill and billing account
https://eaflood.atlassian.net/browse/WATER-4410

> Part of a series of changes to add support for removing a bill licence back into the billing views

We're not ready to implement removing a bill licence from a bill run as it depends on sending requests to the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) and running the legacy refresh job to update everything on our side.

But we can migrate the confirmation page a user sees when removing a licence. Before we get there though we need to add a relationship that has always existed but only thanks to us switching to view based models is available to use.

This updates the `BillModel` and `BillingAccountModel` with relationships that join them We will then be able to extract both bill run, bill and bill licence information in a single [Objection.js query](https://vincit.github.io/objection.js/).