Migrate confirm remove licence from bill run page
https://eaflood.atlassian.net/browse/WATER-4410

> Part of a series of changes to add support for removing a bill licence back into the billing views

We're not ready to implement removing a bill licence from a bill run as it depends on sending requests to the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) and running the legacy refresh job to update everything on our side.

But we can migrate the confirmation page a user sees when removing a licence. Not only is that one less thing in the legacy code but we can also update the view to be consistent with others we've done like

- [Migrate view errored bill run page](https://github.com/DEFRA/water-abstraction-system/pull/785)
- [Migrate view empty bill run page](https://github.com/DEFRA/water-abstraction-system/pull/783)
- [Add cancel bill run page](https://github.com/DEFRA/water-abstraction-system/pull/780)

With this in place we can then fix our billing views and make the [Remove licence button](https://github.com/DEFRA/water-abstraction-system/pull/824) visible again!