Add new Charging Module WaitForStatusRequest (#829)
https://eaflood.atlassian.net/browse/WATER-4390

> Part of a series of changes to migrate the legacy [send a bill run](https://defra.github.io/sroc-charging-module-api-docs/#/bill-run/SendBillRun) to this project

The [Charging Module API](https://github.com/DEFRA/sroc-charging-module.api) generates transaction references for every bill in a bill run and creates the transaction files when it receives a `/send` request. It will respond with an immediate `204` but the process can take some seconds for the larger bill runs. So, clients should not expect a bill run to have a status of `billed` or `billing_not_required` immediately.

Because of this clients need to wait for the status to change from `processing`. We're already doing this differently in our rebill a bill run code. And, when we finally migrate the refreshing a bill run after the CHA [generates it](https://defra.github.io/sroc-charging-module-api-docs/#/bill-run/GenerateBillRun) we'll need this again.

So, we're using this opportunity to create a `WaitForStatusRequest` that should make things easier for ourselves and avoid any duplication in future.