Add new Charging Module ViewBillRunRequest (#830)
https://eaflood.atlassian.net/browse/WATER-4390

> Part of a series of changes to migrate the legacy [send a bill run](https://defra.github.io/sroc-charging-module-api-docs/#/bill-run/SendBillRun) to this project

The [Charging Module API](https://github.com/DEFRA/sroc-charging-module.api) generates transaction references for every bill in a bill run and creates the transaction files when it receives a `/send` request.

Once the process is complete we need to extract those transaction references and apply them to our bill records. When you [view a bill run](https://defra.github.io/sroc-charging-module-api-docs/#/bill-run/ViewBillRun) the top-level details for each bill are returned including the transaction reference. So, if we can make a view bill run request we can get the transaction references we need.

This change adds the new `ViewBillRunRequest`.