Add new Charging Module SendBillRunRequest (#831)
https://eaflood.atlassian.net/browse/WATER-4390

> Part of a series of changes to migrate the legacy [send a bill run](https://defra.github.io/sroc-charging-module-api-docs/#/bill-run/SendBillRun) to this project

The [Charging Module API](https://github.com/DEFRA/sroc-charging-module.api) generates transaction references for every bill in a bill run and creates the transaction files when it receives a `/send` request.

For that to happen though we need to send the 'send a bill' request. However, something that is a button click on our side is 3 stages in the CHA.

- a bill run has to be 'approved' before it can be sent
- the `/send` request is then made
- wait for the CHA to complete the 'send' process and mark the bill run as either `billed` or `billing_not_required`

When we make the request on our side, we don't care about the interstitial stages. So, we are adding a `SendBillRunRequest` which manages all this for us.