fix alignment issue with abstraction condition link (#822)
https://eaflood.atlassian.net/browse/WATER-4332

The link for the abstraction conditions was not on a new line for some of the scenarios. the fix was easy and I applied it to the other sections as well to keep it standardised. also noticed that the billing region should be the last entry so put that at the bottom of the file.