Add missed remove licence button (#824)
https://eaflood.atlassian.net/browse/WATER-4410

When we re-built the view bill page we didn't spot that the old page includes a button that allows you to remove a licence from the bill.

Doh!

The functionality for doing this is still there in the legacy code. We just need to add the missing button back into our view to allow users to access it. There is some logic involved; you should only see the button if the bill run is `ready` and there is more than one licence on the bill.

This change gets the missed button back in.