Add Legacy DeleteBillRequest (#838)
https://eaflood.atlassian.net/browse/WATER-4416

> Part of a series of changes to fix removing a bill not flagging licences for supplementary billing

We're not ready to implement removing a bill from a bill run as it depends on sending requests to the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) and running the legacy refresh job to update everything on our side.

But to trigger all that we need to send a DELETE request to [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service). This adds the legacy request that will deal with sending that request.