Migrate confirm remove bill from bill run page (#837)
https://eaflood.atlassian.net/browse/WATER-4416

We'd recently spotted that the new billing views we'd developed had dropped a 'Remove licence' button. So, we used that opportunity to [Migrate confirm remove licence from bill run page](https://github.com/DEFRA/water-abstraction-system/pull/828) and [added the button back in](https://github.com/DEFRA/water-abstraction-system/pull/824).

But then our conscientious QA team started testing it and found it didn't do what it claimed to do. That was to flag the removed licence for the next supplementary bill run. We display the same message on the confirm remove bill from bill run so now doubtful they tested that as well with the same result.

We've [fixed the remove licence issue](https://github.com/DEFRA/water-abstraction-service/pull/2464). But whilst we fix the legacy code for remove bill run this is a great opportunity to also migrate the confirm remove bill from bill run page. Not only is it one less thing in the legacy code but we can also update the view to be consistent with others we've done like

- [Migrate confirm remove licence from bill run page](https://github.com/DEFRA/water-abstraction-system/pull/828)
- [Migrate view errored bill run page](https://github.com/DEFRA/water-abstraction-system/pull/785)
- [Migrate view empty bill run page](https://github.com/DEFRA/water-abstraction-system/pull/783)
- [Add cancel bill run page](https://github.com/DEFRA/water-abstraction-system/pull/780)