Replace live bill run checking in engine
https://eaflood.atlassian.net/browse/WATER-4416
https://eaflood.atlassian.net/browse/WATER-4379

We spotted an issue with removing a bill from a bill run not setting the supplementary flags on a the licences involved correctly. We've fixed that, but its led to our QA team rigorously re-creating bill runs over and over.

Doing this they spotted we did have an issue with the `CheckLiveBillRunService` that we fixed in [Fix broken CheckLiveBillRunService](https://github.com/DEFRA/water-abstraction-system/pull/841).

Now they've spotted something else. `CheckLiveBillRunService` includes batch type as a filter. This means it _will_ allow you to create a supplementary bill run, for example, even if another type of bill run is 'in progress' (queued, processing, or ready). AT the time we thought that was ok but now we know better.

We had actually spotted this in our work to migrate the setup bill run journey (see [Handle bill run setup matches an existing bill run](https://github.com/DEFRA/water-abstraction-system/pull/810)). We hoped we'd be using our version of the journey by now and we could quietly retire `CheckLiveBillRunService` as part of ongoing maintenance and no one would be any the wiser (it has been live for 7 months now!)

But our QA team would rather clear the issues currently found before bringing in the new setup journey for testing.

So, rather than fix the service, we'll replace it with `DetermineBlockingBillRunService` which matches what the legacy service does and deals with this scenario.