Fix FetchLiveBillRunsService
https://eaflood.atlassian.net/browse/WATER-4375

We migrated the create bill run process from the legacy service as part applying updates to support changes for two-part tariff.

One of the things we had to mirror was how it checks for matching or live bill runs at the end of the journey. In testing we've found there is a scenario our unit tests and therefore the code didn't cover.

> There is _only_ a PRESROC bill run that is `READY` with a financial end year of 2022. You are trying to create an annual bill run for the current financial year (2024).

With that scenario the first test is this and the result would be false.

```javascript
  if (!supplementary && liveBillRun.toFinancialYearEnding === financialYearEnding) {
    matches.push(liveBillRun)
    break
  }
```

Our mistake is to assume only supplementary bill runs would get past this point. So, the next test is passing causing a match to be returned when it shouldn't.

```javascript
  // Because of the check above this will only be applied to supplementary bill runs. This is where we check the
  // live bill run against both the specified financial year ending and the last year of PRESROC 2022
  if (liveBillRun.toFinancialYearEnding === financialYearEnding || liveBillRun.toFinancialYearEnding === LAST_PRESROC_YEAR) {
    matches.push(liveBillRun)
  }
```

This change updates the unit tests to include this scenario and then fixes the flaw in our logic.