Fix PRESROC supplementary flags not clearing
https://eaflood.atlassian.net/browse/WATER-4390

We recently [Migrate legacy send bill run functionality](https://github.com/DEFRA/water-abstraction-system/pull/771) to our system.

The PR details the reasons why (😱🤬) but suffice to say we are now responsible for the final stages of a bill run.

We've been putting the change through its paces and our ever vigilant QA team have spotted something.

When an SROC supplementary bill run is sent the supplementary billing flags on the licences in the bill get cleared. But the same cannot be said for PRESROC supplementary bills.

Clearly there is something we have overlooked that the legacy code was doing that we're not. This change corrects the issue and gets those flags cleared!