Fix supplementary flags not clearing after 'send' (#849)
https://eaflood.atlassian.net/browse/WATER-4390

We recently [Migrate legacy send bill run functionality](https://github.com/DEFRA/water-abstraction-system/pull/771) to our system.

The PR details the reasons why (😱🤬) but suffice to say we are now responsible for the final stages of a bill run.

We've been putting the change through its paces and our ever-vigilant QA team has spotted something.

Once a supplementary bill run is sent the supplementary billing flags on the licences in the bill are supposed to be cleared. But this isn't happening in our version.

After double-checking we've found where the legacy code was doing this. We need to replicate this behaviour in our version!