Do not unflag all SROC supplementary (#852)
https://eaflood.atlassian.net/browse/WATER-4390

We recently [Migrate legacy send bill run functionality](https://github.com/DEFRA/water-abstraction-system/pull/771) to our system.

The PR details the reasons why (😱🤬) but suffice to say we are now responsible for the final stages of a bill run.

We then spotted that we'd overlooked dealing with unflagging licences that were flagged for supplementary billing. That got sorted in [Fix supplementary flags not clearing after 'send'](https://github.com/DEFRA/water-abstraction-system/pull/849).

But then we started discussing the differences between the legacy approach (which we had to recreate) and our approach. We realised there are scenarios we are not covering.

1. User makes a change that flags a licence for supplementary billing
2. User generates a supplementary bill run
3. User then makes another change (adds a new charge version) which is waiting to be approved
4. User sends the bill run

Currently, we'd remove the flag when the licence needs to be included in the next one. Also, if in step 3 they added a new charge version and then approved it, we'd have no workflow record to hang off. We'd be removing the flag when again the licence needs to be in the next supplementary bill run.

This applies to licences the engine determines are unbilled (these get unflagged during bill run creation) and those that were billed (dealt with when the bill run gets sent).

So, this change adds some of that legacy logic to our services to cover these scenarios.