Add service to fetch the abstraction conditions
We were able to build on the fantastic insight that the information we need is in tables so we can avoid the JSON blob. We were able to put together a query that will extract the abstraction conditions within the query rather than having to bring the data back and manipulate the results.

We have done something we don't normally do; process the results within the service. After building the service and passing the raw results to the presenter it felt a better fit to move the processing to the service. That way the knowledge that we have to deal with the situation of the same condition being assigned to different purposes stays in one place.

Also makes testing a lot easier! :-)