Add unit tests
Note the change in the SubmitSendService.

We have been working under the impression that we can return early from a async function by not awaiting the final function.

Whilst in theory this does appear to work its probably not how we should be doing it. This is because for this service (SubmitCancelService is the other place we are following this pattern) we are purposefully throw an error assuming the try/catch in  the controller will see it and log it.

When it came to trying to write a unit test for it we found we just kept blowing up [Hapi/Lab](https://hapi.dev/module/lab/). So, we then tried it for real, forcing `ChargingModuleViewBillRunRequest` to throw an error. We confirmed the controller's try/catch doesn't see the error and we cause 'system' to crash.

The only way to catch the error is to chain `.catch()` directly onto the call. We were only intending to log the error so what we're doing is now fine. But it is worth noting and we may revisit our foreground/background pattern.