Add missing SubmitSendBillRunService unit tests (#851)
https://eaflood.atlassian.net/browse/WATER-4390

In [Migrate legacy send bill run functionality](https://github.com/DEFRA/water-abstraction-system/pull/771) we replaced the mad-as-a-box-of-frogs legacy send functionality with our own!

Only our supposed leader of tech and keeper of quality @Cruikshanks completely missed adding unit tests for the primary `SubmitSendBillRunService`! 🤦

This change adds the missing tests.

---

Note the change in the `SubmitSendService`.

We have been working under the impression that we can return early from an async function by not awaiting the final function.

Whilst in theory this does appear to work it's probably not how we should be doing it. This is because for this service (`SubmitCancelService` is the other place we are following this pattern) we are purposefully throwing an error assuming the `try/catch` in the controller will see it and log it.

When it came to trying to write a unit test for it we found we just kept blowing up [Hapi/Lab](https://hapi.dev/module/lab/). So, we then tried it for real, forcing `ChargingModuleViewBillRunRequest` to throw an error. We confirmed the controller's try/catch doesn't see the error and we caused 'system' to crash.

The only way to catch the error is to chain `.catch()` directly onto the call. We were only intending to log the error so what we're doing is now fine. But it is worth noting and we may need to revisit our foreground/background pattern.