Move transaction matching logic to GeneralLib
https://eaflood.atlassian.net/browse/WATER-4403

We are currently working on updating the supplementary billing engine to incorporate improvements we made to the process when implementing annual billing.

One of the things we've been reminded off is that we have multiple services that use the same logic to determine if the charging information for 2 transactions is the same. It's a bit of a scary matching function which we flagged as a TODO to bring into a single place.

Now seems a good time to do just that and help simplify the supplementary process.