Remove generate bill and bill licence services
https://eaflood.atlassian.net/browse/WATER-4403

We are currently working on updating the supplementary billing engine to incorporate improvements we made to the process when implementing annual billing.

One of the key improvements was fetching the data we need at the billing account level. This means we have a one-to-one relationship between each billing account returned and the bill being created.

When we built the supplementary billing engine this wasn't possible because billing accounts sat in one schema and their charge versions another. We were also fresh to the problem and zeroed in on charge versions being the primary source of data we needed.

So, we ended up building services to generate bills and bill licences as part of a pre-generation logic to try and help get from charge versions to bills. It works, but coming back to it 6 months later and it's not the easiest thing to get back into. Plus we don't need it any more!

Ripping out `app/services/bill-runs/supplementary/pre-generate-billing-data.service.js` is going to need to be a PR in itself. But prior to that we can simplify things by removing `GenerateBillService` and `GenerateBillLicenceService`. It's only `PreGenerateBillingDataService` and the siloed reissue services that use them.